[WIP] Splite teamvite into modules.
Instead of having everything in the same package, split up the logic.

Modules:
- http: anything related to sending or receiving http traffic
- sqlite: anything that interacts with the sqlite database
- teamvite: anything specific to the teamvite app, mostly domain objects and interfaces

This follows recommendations from Ben Johnson on how to organize Go modules and
Ted Kamiski on how to use modules to hide implementation details.

This is the first commit that does a majority of the restructuring and implements
the Game logic (in sqlite, http and teamvite)